Move flood meta data queries into a separate module.
